[LC-41] dispatcher separate into default(node), v2, v3
