+++ iconrpcserver/dispatcher/default/__init__.py
@@ -13,6 +13,3 @@
 # limitations under the License.
 
 from .node import *
-from .version2 import *
-from .version2 import *
-from .version3 import *
