+++ iconrpcserver/dispatcher/v2/__init__.py
@@ -12,5 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .exception import *
-from .dispatcher import *
+from .version2 import *
